src/requirements.txt
discord
selenium
python-dotenv 